feat: add hasheq markup mode
Provides a new markup mode to VOoM where headings are specified using a
combination of hash and equals signs, e.g.:

    # = A top-level heading
    # == A second-level heading
    # === A third-level heading

This new mode was created to provide a method for structuring the
contents of Python files.  The existing `hashes` mode was not considered
to be acceptable because it caused the flake8 static analysis tool to
raise the error `E266 too many leading '#' for block comment` each time
a hash-based heading was encountered.

The format was chosen to mimic the appearance of MediaWiki heading
markup, and unlike the `hashes` mode, at least one space character is
required between the final equals sign and the heading text.